refactor(net): extract isIP, isIPv4, and isIPv6
